pyproject.toml
[tool.poetry]
name = "locust-live"
version = "0.0.1"
description = ""
authors = ["dunossauro <mendesxeduardo@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.7"
flask = "^1.1"
flask-login = "^0.4.1"
Flask-SQLAlchemy = "^2.4"
flask-migrate = "^2.5"

[tool.poetry.dev-dependencies]
factory_boy = "^2.12"
tox = "^3.14"
coverage = "^4.5"
pycodestyle = "^2.5"
pydocstyle = "^4.0"
pyflakes = "^2.1"
pylava = "^0.2.2"
xenon = "^0.7.0"
mypy = "^0.740.0"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
